Moved About Us and How it works on a different page
