+++ frontend/src/App.js
@@ -17,6 +17,8 @@ import LogoutPage from "./pages/Auth_Pages/LogoutPage";
 import FriendsPage from "./pages/friends/FriendsPage";
 import ProfilePage from "./pages/profile/ProfilePage";
 import AccountPage from "./pages/account/AccountPage";
+import AboutUs from "./pages/getting_started/AboutUs";
+import HowItWorks from "./pages/getting_started/HowItWorks";
 
 // import nav components
 import NavBar from "./components/NavBar/NavBar";
@@ -96,6 +98,8 @@ function AppContent() {
           <Route path="/profile" element={<ProfilePage />} />
           <Route path="/friends" element={<FriendsPage />} />
           <Route path="/logout" element={<LogoutPage />} />
+          <Route path="/about-us" element={<AboutUs/>} />
+          <Route path="/how-it-works" element={<HowItWorks />} />
         </Routes>
 			</div>
 	);
