+++ frontend/src/pages/getting_started/AboutUs.js
@@ -1,0 +1,40 @@
+import React from 'react'
+import "../../css/GettingStarted.css"
+function AboutUs() {
+  return (
+    <div className='about-us-container'>
+      <div>
+       
+      </div>
+
+      {/* About RightWay */}
+      <div>
+        <div className="banner">
+          <img src="/Highland.jpg" alt="" />
+
+          <div className="about-text-section">
+            <div>
+              <h1>About us</h1>
+            </div>
+            <div>
+              <p>
+                {" "}
+                At RightWay, we believe that the journey is just as important as
+                the destination. That's why we've created a platform designed to
+                bring friends together through unforgettable travel experiences,
+                tailored just for you. Whether you're planning a weekend getaway
+                or an adventure of a lifetime, RightWay takes the hassle out of
+                trip planning.
+              </p>
+            </div>
+          </div>
+        </div>
+      </div>
+    
+    </div>
+   
+    
+  )
+}
+
+export default AboutUs
